+++ lib/glimmer/dsl/widget_listener_expression.rb
@@ -6,17 +6,19 @@ module Glimmer
       include_package 'org.eclipse.swt.widgets'
 
       def can_interpret?(parent, keyword, *args, &block)
-        Glimmer.logger.debug "keyword starts with on_: #{keyword.start_with?('on_')}"
-        return unless keyword.start_with?('on_')
-        Glimmer.logger.debug "block exists?: #{!block.nil?}"
-        return unless !block.nil?
+        Glimmer.logger&.debug "keyword starts with on_: #{keyword.start_with?('on_')}"
+        return false unless keyword.start_with?('on_')
         widget_parentage = widget?(parent)
-        Glimmer.logger.debug "parent is a widget: #{widget_parentage}"
-        return unless widget_parentage
-        Glimmer.logger.debug "args are empty?: #{args.empty?}"
-        return unless args.empty?
+        Glimmer.logger&.debug "parent is a widget: #{widget_parentage}"
+        return false unless widget_parentage
+        Glimmer.logger&.debug "block exists?: #{!block.nil?}"
+        raise Glimmer::Error, "Listener is missing block for keyword: #{keyword}" unless block_given?
+        Glimmer.logger&.debug "args are empty?: #{args.empty?}"
+        raise Glimmer::Error, "Invalid listener arguments for keyword: #{keyword}(#{args.inspect})" unless args.empty?
         result = parent.can_handle_observation_request?(keyword)
-        Glimmer.logger.debug "can add listener? #{result}"
+        Glimmer.logger&.debug "can add listener? #{result}"
+        raise Glimmer::Error, "Invalid listener keyword: #{keyword}" unless result
+        true
       end
 
       def interpret(parent, keyword, *args, &block)