+++ lib/glimmer/swt/layout_data_proxy.rb
@@ -31,7 +31,7 @@ module Glimmer
         begin
           @swt_layout_data = swt_layout_data_class.new(*args)
         rescue => e
-          Glimmer.logger.debug "#{e.message}\n#{e.backtrace.join("\n")}"
+          Glimmer.logger&.debug "#{e.message}\n#{e.backtrace.join("\n")}"
           @swt_layout_data = args.first if args.count == 1
         end
         @widget_proxy.swt_widget.setLayoutData(@swt_layout_data)
@@ -62,11 +62,22 @@ module Glimmer
 
       def set_attribute(attribute_name, *args)
         args = SWTProxy.constantify_args(args)
-        @swt_layout_data.send(attribute_setter(attribute_name), *args)
+        if args.first != @swt_layout_data.send(attribute_getter(attribute_name))
+          @swt_layout_data.send(attribute_setter(attribute_name), *args)
+          @widget_proxy.swt_widget.getShell.pack
+        end
+      end
+
+      def get_attribute(attribute_name)
+        @swt_layout_data.send(attribute_getter(attribute_name))
       end
 
       def attribute_setter(attribute_name)
         "#{attribute_name.to_s.camelcase(:lower)}="
+      end
+
+      def attribute_getter(attribute_name)
+        "#{attribute_name.to_s.camelcase(:lower)}"
       end
     end
   end